requirements.txt
TikTokApi>=6.0.0
playwright>=1.40.0
requests>=2.31.0
python-dotenv>=1.0.0
sqlalchemy>=2.0.0
selenium>=4.15.0
webdriver-manager>=4.0.1
fake-useragent>=1.4.0
yt-dlp>=2024.0.0  # Téléchargement TikTok avec gestion des tokens

